+++ src/evals/scenario/ai-playwright.ts
@@ -3,22 +3,21 @@ import * as path from "path";
 import { DESCRIPTIONS } from "../../server.js";
 import { getHumeMcpTools } from "../utils.js";
 import {
-  EvalScenario,
   commonCriteria,
   commonInstructions,
   voiceDesignCriteria,
-} from "./types.js";
+} from "./common.js";
 import {
   getContent,
   handler,
   mockDisplayResult,
   mockDisplayUse,
-} from "./helpers.js";
+} from "./mock.js";
+import { EvalScenario } from "../roleplay.js";
 
 export const aiPlaywrightScenario = async (
   descriptions: typeof DESCRIPTIONS,
 ): Promise<EvalScenario> => {
-  // Read dialogue content from data file
   const dialogueContent = await fs.readFile(
     path.join(__dirname, "/../data/play_dialogue.txt"),
     "utf-8",
